backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.20
langchain-openai==0.1.6
langchain-core==0.1.52
pydantic==1.10.13
python-dotenv==1.0.0
httpx==0.27.0
